backend-prod/requirements.txt
Click==7.0
dnspython==1.16.0
eventlet==0.25.1
Flask==1.1.1
Flask-SocketIO==4.2.1
greenlet==0.4.15
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
monotonic==1.5
pylsl==1.12.2
python-engineio==3.11.2
python-socketio==4.4.0
six==1.14.0
Werkzeug==1.0.0
aiohttp
async-timeout
chardet
multidict
yarl 